fixing to_string

+++ BayesClassifier/main.cpp
@@ -212,7 +212,7 @@ int main(int argc, char* argv[]) {
 
 		cout << "probabilities" << endl;
 		for (map<string, double>::iterator category = category_probs.begin(); category != category_probs.end(); category++) {
-			cout << category->first << ":" << to_string(category->second / prob_sum) << endl;
+			cout << category->first << ":" << to_string((double long)category->second / prob_sum) << endl;
 		}
 
 		cout << "prediction: " << prediction << endl;
@@ -226,5 +226,5 @@ int main(int argc, char* argv[]) {
 		}
 	}
 	double accuracy = double(correct_count) / double(test_set.size());
-	cout << "Accuracy: " << to_string(accuracy) << endl;
+	cout << "Accuracy: " << to_string((double long)accuracy) << endl;
 }
